// src/generated/models/Mserp_hcmcandidatetohireentitiesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Mserp_hcmcandidatetohireentitiesmserp_applicantintegrationresult = {
  200000000: 'NotProcessed',
  200000001: 'Hired',
  200000002: 'Nothired',
  200000003: 'Dismissed'
} as const;
export type Mserp_hcmcandidatetohireentitiesmserp_applicantintegrationresult = keyof typeof Mserp_hcmcandidatetohireentitiesmserp_applicantintegrationresult;
export const Mserp_hcmcandidatetohireentitiesmserp_gender = {
  200000000: 'None',
  200000001: 'Male',
  200000002: 'Female',
  200000003: 'Non_Specific'
} as const;
export type Mserp_hcmcandidatetohireentitiesmserp_gender = keyof typeof Mserp_hcmcandidatetohireentitiesmserp_gender;
export const Mserp_hcmcandidatetohireentitiesmserp_isdisabledveteran = {
  200000000: 'No',
  200000001: 'Yes'
} as const;
export type Mserp_hcmcandidatetohireentitiesmserp_isdisabledveteran = keyof typeof Mserp_hcmcandidatetohireentitiesmserp_isdisabledveteran;
export const Mserp_hcmcandidatetohireentitiesmserp_iswillingtorelocate = {
  200000000: 'No',
  200000001: 'Yes'
} as const;
export type Mserp_hcmcandidatetohireentitiesmserp_iswillingtorelocate = keyof typeof Mserp_hcmcandidatetohireentitiesmserp_iswillingtorelocate;

export interface Mserp_hcmcandidatetohireentitiesBase {
  mserp_applicantintegrationresult?: Mserp_hcmcandidatetohireentitiesmserp_applicantintegrationresult;
  mserp_availabilitydate?: string;
  mserp_birthdate?: string;
  mserp_candidateid?: string;
  mserp_citizenshipcountrycode?: string;
  mserp_comments?: string;
  mserp_dataareaid?: string;
  "mserp_dataAreaId_id@odata.bind"?: string;
  mserp_donothirereasoncodeid?: string;
  mserp_ethnicoriginid?: string;
  mserp_firstname?: string;
  "mserp_FK_Position_id@odata.bind"?: string;
  mserp_gender?: Mserp_hcmcandidatetohireentitiesmserp_gender;
  mserp_hcmcandidatetohireentityid: string;
  mserp_isdisabledveteran?: Mserp_hcmcandidatetohireentitiesmserp_isdisabledveteran;
  mserp_iswillingtorelocate?: Mserp_hcmcandidatetohireentitiesmserp_iswillingtorelocate;
  mserp_lastname?: string;
  mserp_lastnameprefix?: string;
  mserp_middlename?: string;
  mserp_militaryserviceenddate?: string;
  mserp_militaryservicestartdate?: string;
  mserp_partynumber?: string;
  mserp_positionid?: string;
  mserp_recruitingrequestid?: string;
  mserp_veteranstatusid?: string;
}

export interface Mserp_hcmcandidatetohireentities extends Mserp_hcmcandidatetohireentitiesBase {
  mserp_applicantintegrationresultname?: string;
  mserp_dataareaid_idname?: string;
  mserp_fk_position_idname?: string;
  mserp_gendername?: string;
  mserp_isdisabledveteranname?: string;
  mserp_iswillingtorelocatename?: string;
  mserp_partytype?: string;
  mserp_dataareaid_id?: object;
  _mserp_dataareaid_id_value?: string;
  mserp_fk_position_id?: object;
  _mserp_fk_position_id_value?: string;
}